backend/pyproject.toml
[project]
name = "reddit-mvp-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.2",
    "django==5.2",
    "django-ninja>=1.5.3",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "praw>=7.8.1",
    "psycopg[binary]>=3.3.2",
    "redis>=7.1.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.13",
]
